gameui cMenuRun/pauseMenuRunSubmenu: (int) cast on gCMenuButtons bit tests -> signed cmpwi (#14)
cMenuRun 98.13->98.23, pauseMenuRunSubmenu 96.32->96.50. gCMenuButtons is
extern u32, so (g & 0x200)==0 / (g & 0x100) emitted cmplwi; target uses signed
cmpwi after the rlwinm single-bit extract. (int) cast forces the signed compare,
matching the established (int)btn & 0x200 idiom already at line 1336.

(cherry picked from commit a1706104f41366706138affc8c3ff1f433e7d5a3)

+++ src/main/dll/dll_0000_gameui.c
@@ -1881,7 +1881,7 @@ void pauseMenuRunSubmenu(u8 p1)
                 valid = 1;
             }
         }
-        if ((gCMenuButtons & 0x100) && tbl != lbl_8031BD30 && 0.0f == lbl_803DD7C0)
+        if (((int)gCMenuButtons & 0x100) && tbl != lbl_8031BD30 && 0.0f == lbl_803DD7C0)
         {
             if (valid != 0)
             {
@@ -2171,7 +2171,7 @@ void cMenuRun(void)
                         }
                     }
                 }
-                else if (gCMenuButtons & 0x200)
+                else if ((int)gCMenuButtons & 0x200)
                 {
                     Sfx_PlayFromObject(0, 0x37c);
                     cMenuOpen = 0;
